Refactor inside Vectorization.
Bench: 2886048
