+++ Backend/Engine/NNUE/Vectorization/Intrinsic.cs
@@ -3,9 +3,8 @@ using System.Numerics;
 using System.Runtime.CompilerServices;
 using System.Runtime.Intrinsics;
 using System.Runtime.Intrinsics.X86;
-using Backend.Engine.NNUE.Vectorization;
 
-namespace Backend.Engine.NNUE;
+namespace Backend.Engine.NNUE.Vectorization;
 
 public static class Intrinsic
 {
