+++ Backend/Engine/NNUE/Vectorization/NN.cs
@@ -1,9 +1,8 @@
 ﻿using System.Numerics;
 using System.Runtime.CompilerServices;
 using Backend.Data.Enum;
-using Backend.Engine.NNUE.Vectorization;
 
-namespace Backend.Engine.NNUE;
+namespace Backend.Engine.NNUE.Vectorization;
 
 public static class NN
 {
